Partial refactor to support regular expression backreferences for replacements. uap-core has started to use these for UAParser and OSParser instances, and only DeviceParser currently supports.
The spec (https://github.com/ua-parser/uap-core/blob/master/docs/specification.md)
indicates that all matches from $1 to $9 should be honored for every
replacement.

It also allows for a v3 replacement for UserAgent (which is used for at least one
useragent at this point in time), so I added that as well.

Also refactored the class hierarchy so that it's sort of DRY,
and moved MultiReplace to a module level fuction because it doesn't use
self.

Bumped version to 0.7.4